requirements.txt
discord.py==1.3.4
requests==2.24.0
beautifulsoup4==4.7.1
firebase-admin==5.0.3
pycryptodome==3.10.1